update 예약 및 결제

+++ src/main/java/com/sh/pettopia/choipetsitter/controller/PetSitterController.java
@@ -58,12 +58,15 @@ public class PetSitterController {
 
         if (!checkFiles.isEmpty()) {
             // 기존에 클라우드와 db에 저장된 파일을 삭제한다.
-            String url=petSitter.getMainImageUrl(); // 전체 URL이기 때문에 삭제할 때는 파일이름이 필요하다
-            String keyword = "mainImage/";
-            int startIndex = url.indexOf(keyword) + keyword.length();
-            String mainImageFileName = url.substring(startIndex);
-            fileService.deleteImage(petSitter.getPetSitterId(), "mainImage", mainImageFileName);
 
+            String url=petSitter.getMainImageUrl(); // 전체 URL이기 때문에 삭제할 때는 파일이름이 필요하다
+            if(url!=null) {
+                String keyword = "mainImage/";
+                int startIndex = url.indexOf(keyword) + keyword.length();
+                String mainImageFileName = url.substring(startIndex);
+                fileService.deleteImage(petSitter.getPetSitterId(), "mainImage", mainImageFileName);
+                log.info("mainImageFileName = {}", mainImageFileName);
+            }
 
             //파일들의 url을 List로 담아서 Dto->Entity할 때 넣는다
             List<FileDto> fileDtoList = fileService.sitterUpFile(checkFiles, principal.getMember().getEmail(), "mainImage");
@@ -87,7 +90,6 @@ public class PetSitterController {
 
         petSitterService.save(petSitter);
         System.out.println("petSitter = " + petSitter);
-        redirectAttributes.addFlashAttribute("message", "프로필을 성공적으로 등록 했습니다");
 
         return "redirect:/petsitter/registerprofile";
     }
@@ -105,7 +107,7 @@ public class PetSitterController {
     public String detail(@PathVariable String petSitterId, Model model) {
         log.info("GET /petsitter/detail/{}", petSitterId);
         PetSitter petSitter = petSitterService.findOneByPetSitter(petSitterId);
-        PetSitterRegisterDto dto = new PetSitterRegisterDto().EntityToDto(petSitter);
+        PetSitterRegisterDto dto = new PetSitterRegisterDto().entityToDto(petSitter);
         List<PetSitterReview> PetSitterReviewList = petSitterReviewService.findPetSitterReviewByPetSitterId(petSitterId);
 
         List<PetSitterReviewDto> petSitterReviewDtoList = new ArrayList<>();
@@ -171,20 +173,17 @@ public class PetSitterController {
             petSizeList.add(petSize.getPetSize());
         }
 
-        System.out.println("stringService = " + petServiceList);
-        System.out.println("petSizeList = " + petSizeList);
-        System.out.println("petSitter = " + petSitter);
+        log.info("stringService = {}" , petServiceList);
+        log.info("petSizeList = {}" , petSizeList);
+        log.info("petSitter = {}" , petSitter);
 
-        if (petSitter != null) {
-
-            PetSitterRegisterDto dto = new PetSitterRegisterDto().EntityToDto(petSitter);
+            PetSitterRegisterDto dto = new PetSitterRegisterDto().entityToDto(petSitter);
             dto.setPostImagesList(profileImg);
             model.addAttribute("dto", dto);
             model.addAttribute("petSize", petSizeList);
             model.addAttribute("petService", petServiceList);
             // 여기서 항상 서버에서 이미지 url을 담는다 -> html에서 그 url이 입력되면 이미지는 출력이된다
 
-        }
         // 펫시터 홍보글 작성
     }
 
@@ -233,7 +232,6 @@ public class PetSitterController {
         log.info("GET /register_profile");
         log.debug("principal = {}", principal.getMember().getEmail());
 
-        List<String> mainImageUrl = new ArrayList<>(); // NCP 서버에 등록된 주소를 가져오기 위한 객체
         Set<String> stringService = new HashSet<>(); // Set<enum>은 html에서 contains가 안될 때가 있어서 문자열로 비교하기 위해 따로 담아서 model에 넘긴다
         Set<String> stringPetSize = new HashSet<>();
 
@@ -251,7 +249,7 @@ public class PetSitterController {
         System.out.println("stringPetSize = " + stringPetSize);
         System.out.println("petSitter = " + petSitter);
 
-        PetSitterRegisterDto petSitterRegisterDto = new PetSitterRegisterDto().EntityToDto(petSitter);
+        PetSitterRegisterDto petSitterRegisterDto = new PetSitterRegisterDto().entityToDto(petSitter);
 
         if (petSitter != null) {
             log.info("petSitterRegisterDto = {}",petSitterRegisterDto);
